requirements.txt
Django
djangorestframework==3.11.0
mysql-connector==2.2.9
mysqlclient==1.3.13
django-heroku == 0.3.1
gunicorn==19.9.0
whitenoise==4.1.2
django-cors-headers==3.4.0




